Subarrays with product less than k
Either I am doing this incorrectly or the solution is incorrect, the given solution says that it is O(n^3) because of generating subarrays for each loop iteration, but this doesn't seem right.

They seem to be trying to add all subarrays of the window within a single go which seems unnecessary.