Impl invert binary tree (kinda)
At its current state, this does not work (at least the multiprocessed version). This is because it's non-trivial to modify a single binary tree object across multiple processes, and frankly that work is not what's important for my studying.

The possible solutions here to get things working are to redefine a binary tree in terms of managed lists or to create a proxy object for binary trees. There may be other solutions involving a queue or a pipe but again, not relevant.